project-development/statistics-calendar-api/requirements.txt
requests
python-dotenv
matplotlib
Flask
flask-cors